Move verify_answer_access to accesshelper
